+++ src/main/kotlin/lotto/domain/model/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.domain.model
 
 import lotto.Constants
