+++ src/main/kotlin/lotto/domain/model/Rank.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.domain.model
 
 enum class Rank(
     val countOfMatch: Int,
