+++ src/test/kotlin/lotto/model/LottoMachineTest.kt
@@ -1,5 +1,6 @@
 package lotto.model
 
+import lotto.domain.service.LottoMachine
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
