+++ src/test/kotlin/lotto/model/LottoNumberTest.kt
@@ -1,5 +1,6 @@
 package lotto.model
 
+import lotto.domain.model.LottoNumber
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
