+++ src/test/kotlin/lotto/model/LottoResultTest.kt
@@ -1,5 +1,7 @@
 package lotto.model
 
+import lotto.domain.model.Rank
+import lotto.domain.service.LottoResult
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
